put all of the classes that were in main.dart into their own separate files for readability
