+++ .vscode/launch.json
@@ -1,0 +1,25 @@
+{
+    "configurations": [
+        {
+            "name": "Flutter (lib/loginPage.dart)",
+            "type": "dart",
+            "request": "launch",
+            "cwd": "pup_iq",
+            "program": "lib/loginPage.dart"
+        },
+        {
+            "name": "Flutter (lib/main1.dart)",
+            "type": "dart",
+            "request": "launch",
+            "cwd": "pup_iq",
+            "program": "lib/main1.dart"
+        },
+        {
+            "name": "Flutter (lib/main.dart)",
+            "type": "dart",
+            "request": "launch",
+            "cwd": "pup_iq",
+            "program": "lib/main.dart"
+        }
+    ]
+}
