+++ pup_iq/lib/hardPage.dart
@@ -1,0 +1,93 @@
+import 'package:flutter/material.dart';
+import 'main.dart';
+import 'puppy.dart';
+
+class HardPage extends StatelessWidget{
+  Widget build(BuildContext context){
+    return Scaffold(
+      appBar: AppBar(title: Text("Hard Lessons"),
+      backgroundColor: Colors.blue),
+      body: Column(
+          crossAxisAlignment: CrossAxisAlignment.stretch,
+          children: [
+            Expanded(
+              child: ListView(
+                children:[
+                  
+                  GestureDetector(
+                    onTap: () {
+                      Navigator.push(context, MaterialPageRoute(builder: (context)=> CrawlPage()),);
+                    },
+                    child: Container(
+                      height: 205,
+                      color: Color.fromRGBO(33, 150, 243, 1).withOpacity(.3),
+                      child: Center(
+                        child: Text("Crawl",
+                        style: TextStyle(fontSize: 25,
+                        color: Colors.black,
+                        fontWeight: FontWeight.bold,
+                        ),
+                        ),
+                      ),
+                      ),
+                  ),
+                  GestureDetector(
+                    onTap: () {
+                    },
+                    child: Container(
+                      height: 205,
+                      color: Color.fromRGBO(33, 150, 243, 1).withOpacity(.2),
+                      child: Center(
+                        child: Text("Spin/Not Done",
+                        style: TextStyle(fontSize: 25,
+                        color: Colors.red,
+                        fontWeight: FontWeight.bold,
+                        ),
+                        ),
+                      ),
+                      ),
+                  ),
+                  GestureDetector(
+                    onTap: () {
+                    },
+                    child: Container(
+                      height: 205,
+                      color: Color.fromRGBO(33, 150, 243, 1).withOpacity(.3),
+                      child: Center(
+                        child: Text("Recall/Not Done",
+                        style: TextStyle(fontSize: 25,
+                        color: Colors.red,
+                        fontWeight: FontWeight.bold,
+                        ),
+                        ),
+                      ),
+                      ),
+                  ),
+                  GestureDetector(
+                    onTap: () {
+                      Navigator.push(context, MaterialPageRoute(builder: (context) => BellsPage()),);
+                    },
+                    child: Container(
+                      height: 205,
+                      color: Color.fromRGBO(33, 150, 243, 1).withOpacity(.2),
+                      child: Center(
+                        child: Text("Ring Bell For Bathroom",
+                        style: TextStyle(fontSize: 25,
+                        color: Colors.black,
+                        fontWeight: FontWeight.bold,
+                        ),
+                        ),
+                      ),
+                      ),
+                  ),
+
+
+                ],
+                ),
+            ),
+          ],
+        ),
+    );
+  }
+  
+}
