+++ pup_iq/lib/newDogPage.dart
@@ -1,0 +1,90 @@
+import 'package:flutter/material.dart';
+import 'main.dart';
+import 'puppy.dart';
+
+class NewDogPage extends StatelessWidget{
+  Widget build(BuildContext context){
+    return Scaffold(
+      appBar: AppBar(title: Text("Hard Lessons"),
+      backgroundColor: Colors.blue),
+      body: Column(
+          crossAxisAlignment: CrossAxisAlignment.stretch,
+          children: [
+            Expanded(
+              child: ListView(
+                children:[
+                  
+                  GestureDetector(
+                    onTap: () {
+                    },
+                    child: Container(
+                      height: 205,
+                      color: Color.fromRGBO(33, 150, 243, 1).withOpacity(.3),
+                      child: Center(
+                        child: Text("Potty Training",
+                        style: TextStyle(fontSize: 25,
+                        color: Colors.black,
+                        fontWeight: FontWeight.bold,
+                        ),
+                        ),
+                      ),
+                      ),
+                  ),
+                  GestureDetector(
+                    onTap: () {
+                    },
+                    child: Container(
+                      height: 205,
+                      color: Color.fromRGBO(33, 150, 243, 1).withOpacity(.2),
+                      child: Center(
+                        child: Text("Crate Training",
+                        style: TextStyle(fontSize: 25,
+                        color: Colors.black,
+                        fontWeight: FontWeight.bold,
+                        ),
+                        ),
+                      ),
+                      ),
+                  ),
+                  GestureDetector(
+                    onTap: () {
+                    },
+                    child: Container(
+                      height: 205,
+                      color: Color.fromRGBO(33, 150, 243, 1).withOpacity(.3),
+                      child: Center(
+                        child: Text("Name Learning",
+                        style: TextStyle(fontSize: 25,
+                        color: Colors.black,
+                        fontWeight: FontWeight.bold,
+                        ),
+                        ),
+                      ),
+                      ),
+                  ),
+                  GestureDetector(
+                    onTap: () {
+                    },
+                    child: Container(
+                      height: 205,
+                      color: Color.fromRGBO(33, 150, 243, 1).withOpacity(.2),
+                      child: Center(
+                        child: Text("Leash Introduction",
+                        style: TextStyle(fontSize: 25,
+                        color: Colors.black,
+                        fontWeight: FontWeight.bold,
+                        ),
+                        ),
+                      ),
+                      ),
+                  ),
+
+                ],
+                ),
+            ),
+          ],
+        ),
+    );
+  }
+  
+}
