+++ pup_iq/lib/newOwnersPage.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+import 'puppy.dart';
+
+class NewOwnersPage extends StatelessWidget{
+  Widget build(BuildContext context){
+    return Scaffold(
+      appBar: AppBar(title: Text("New Owners"),
+      backgroundColor: Colors.blue),
+    );
+  }
+  
+}
