--- pup_iq/lib/practice.dart
@@ -1,34 +1,0 @@
-import 'package:flutter/material.dart';
-//Lani & Tyler's Project
-
-void main() {
-  runApp(MaterialApp(
-    home: Scaffold(
-      appBar: AppBar(
-        title: Text('pupIQ'),
-        ),
-      ),
-      ));
-}
-
-/*class MyApp extends StatelessWidget {
-  MyApp({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Flutter Demo',
-      home: PurpleBackground(), // Use the PurpleBackground widget as the home
-    );
-  }
-}
-
-class PurpleBackground extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      color: Colors.purple, // Set the background color to purple
-    );
-  }
-}*/
-
