+++ pup_iq/lib/resources.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+import 'package:youtube_player_flutter/youtube_player_flutter.dart';
+import 'profilePage.dart';
+import 'puppy.dart';
+
+class resourcesPage extends StatelessWidget{
+  Widget build(BuildContext context){
+    return Scaffold(
+      appBar: AppBar(title: Text("Resources"),
+      backgroundColor: Colors.blue),
+    );
+  }
+  
+}
