+++ pup_iq/pubspec.yaml
@@ -30,6 +30,7 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
+  shared_preferences: ^2.0.8
 
 
   # The following adds the Cupertino Icons font to your application.
